organize cli, add cli tests
