+++ modules/cookies/cookie-types.ts
@@ -1,0 +1,8 @@
+export type CookieOptions = {
+    maxAge?: number;
+    path?: string;
+    domain?: string;
+    secure?: boolean;
+    httpOnly?: boolean;
+    sameSite?: "Strict" | "Lax" | "None";
+  };
