+++ plugins/react/tsconfig.json
@@ -1,0 +1,31 @@
+{
+    "compilerOptions": {
+      "lib": ["ESNext", "DOM", "DOM.Iterable", "es6", "dom"],
+      "module": "esnext",
+      "target": "esnext",
+      "moduleResolution": "bundler",
+      "moduleDetection": "force",
+      "strict": true,
+      "downlevelIteration": true,
+      "skipLibCheck": true,
+      "jsx": "react-jsx",
+      "allowSyntheticDefaultImports": true,
+      "forceConsistentCasingInFileNames": true,
+      "allowJs": false,
+      "types": [
+        "bun-types" // add Bun global
+      ],
+      "baseUrl": ".",
+      "paths": {
+        "mod/*": ["modules/*"],
+      }
+    },
+    "include": [
+      "./modules/**/*.ts",
+      "modules/**/*.ts",
+      "apps/**/*.ts",
+      "_examples/**/*.ts",
+      "./release.ts"
+    ],
+  }
+
